Fix for SNES lag
Instead of setting the "enabled" key to "false" (which seems to cause tons of lag on the SNES for some reason), this just sets the alpha of all thumbnails to 0. This modification provides the exact same end result as before, whilst getting rid of the lag.

NOTE: I don't have a NES mini to test on, but I can't see why this wouldn't work on there also.